# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vendora API",
        "description": "Vendora personal finance API. Use Bearer token authentication on protected endpoints.",
        "contact": {
            "email": "admin@vendora.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/market/quote": {
            "get": {
                "tags": [
                    "Market Data"
                ],
                "summary": "Get live quote for a symbol",
                "operationId": "e6019b3cd46b8075eee1bf13a4c1d09c",
                "parameters": [
                    {
                        "name": "symbol",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "AAPL"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quote data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/market/search": {
            "get": {
                "tags": [
                    "Market Data"
                ],
                "summary": "Search for market symbols",
                "operationId": "8175f79460e82ed42fbbdf53e1fb4dce",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "apple"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/market/candle": {
            "get": {
                "tags": [
                    "Market Data"
                ],
                "summary": "Get candle/sparkline data for a symbol",
                "operationId": "4faad4cef38a9d19d3fb229e5d4fc4e5",
                "parameters": [
                    {
                        "name": "symbol",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "AAPL"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sparkline data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "List user notifications",
                "operationId": "911565ad9aff9ba7eded3f189509e578",
                "parameters": [
                    {
                        "name": "unread_only",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of notifications"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/unread-count": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get unread notification count",
                "operationId": "0272623bed30dd4d72141741fb3f25da",
                "responses": {
                    "200": {
                        "description": "Unread count"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}/read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark a notification as read",
                "operationId": "65581bb40eba3764fb900f21495814de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/read-all": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark all notifications as read",
                "operationId": "116662a87d4b13956efde9efdfbd18e6",
                "responses": {
                    "200": {
                        "description": "All marked as read"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}": {
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Delete a notification",
                "operationId": "54c4eb5ec5dc0d1e2916cec2fced1056",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}